# src/main/resources/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Product API",
    "description": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
...
    "/api/v1/product/prices/find/all": {
      "get": {
        "tags": [
          "price-controller"
        ],
        "summary": "findAllByProductId",
        "operationId": "findAllByProductIdUsingGET",
        "parameters": [
          {
            "name": "pageNo",
            "in": "query",
            "description": "pageNo",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "productId",
            "in": "query",
            "description": "productId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/product/prices/find/default": {
      "get": {
        "tags": [
          "price-controller"
        ],
        "summary": "getDefaultPriceByProductId",
        "operationId": "getDefaultPriceByProductIdUsingGET",
        "parameters": [
          {
            "name": "priceId",
            "in": "query",
            "description": "priceId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/product/prices/find/one": {
      "get": {
        "tags": [
          "price-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET",
        "parameters": [
          {
            "name": "priceId",
            "in": "query",
            "description": "priceId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResponseEntity": {
        "title": "ResponseEntity",
        "type": "object",
        "properties": {
          "body": {
            "type": "object"
          },
          "statusCode": {
            "type": "string",
            "enum": [
              "ACCEPTED",
              "ALREADY_REPORTED",
              "BAD_GATEWAY",
              "BAD_REQUEST",
              "BANDWIDTH_LIMIT_EXCEEDED",
              "CHECKPOINT",
              "CONFLICT",
              "CONTINUE",
              "CREATED",
              "DESTINATION_LOCKED",
              "EXPECTATION_FAILED",
              "FAILED_DEPENDENCY",
              "FORBIDDEN",
              "FOUND",
              "GATEWAY_TIMEOUT",
              "GONE",
              "HTTP_VERSION_NOT_SUPPORTED",
              "IM_USED",
              "INSUFFICIENT_SPACE_ON_RESOURCE",
              "INSUFFICIENT_STORAGE",
              "INTERNAL_SERVER_ERROR",
              "I_AM_A_TEAPOT",
              "LENGTH_REQUIRED",
              "LOCKED",
              "LOOP_DETECTED",
              "METHOD_FAILURE",
              "METHOD_NOT_ALLOWED",
              "MOVED_PERMANENTLY",
              "MOVED_TEMPORARILY",
              "MULTIPLE_CHOICES",
              "MULTI_STATUS",
              "NETWORK_AUTHENTICATION_REQUIRED",
              "NON_AUTHORITATIVE_INFORMATION",
              "NOT_ACCEPTABLE",
              "NOT_EXTENDED",
              "NOT_FOUND",
              "NOT_IMPLEMENTED",
              "NOT_MODIFIED",
              "NO_CONTENT",
              "OK",
              "PARTIAL_CONTENT",
              "PAYLOAD_TOO_LARGE",
              "PAYMENT_REQUIRED",
              "PERMANENT_REDIRECT",
              "PRECONDITION_FAILED",
              "PRECONDITION_REQUIRED",
              "PROCESSING",
              "PROXY_AUTHENTICATION_REQUIRED",
              "REQUESTED_RANGE_NOT_SATISFIABLE",
              "REQUEST_ENTITY_TOO_LARGE",
              "REQUEST_HEADER_FIELDS_TOO_LARGE",
              "REQUEST_TIMEOUT",
              "REQUEST_URI_TOO_LONG",
              "RESET_CONTENT",
              "SEE_OTHER",
              "SERVICE_UNAVAILABLE",
              "SWITCHING_PROTOCOLS",
              "TEMPORARY_REDIRECT",
              "TOO_EARLY",
              "TOO_MANY_REQUESTS",
              "UNAUTHORIZED",
              "UNAVAILABLE_FOR_LEGAL_REASONS",
              "UNPROCESSABLE_ENTITY",
              "UNSUPPORTED_MEDIA_TYPE",
              "UPGRADE_REQUIRED",
              "URI_TOO_LONG",
              "USE_PROXY",
              "VARIANT_ALSO_NEGOTIATES"
            ]
          },
          "statusCodeValue": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}